Update rust-bitcoin to v0.30.1
Some API JSON data types were updated to match changes in the
underlying rust-bitcoin data types:

- vsize and weights were changed from u32 to u64 (in `GET /tx/:txid`,
  `GET /block/:hash`, `GET /mempool` and `GET /mempool/recent`)

- The `fee_histogram` was changed from [(f32,u32)] to [(f64,u64)]
  (in `GET /mempool` and Electrum's `mempool.get_fee_histogram`)

- Block mining difficulty was changed from u64 to u128
  (in GET `/block/:hash`)

The actual JSON string serialization should remain unchanged, as
serde-json serializes u64/u128/f64 as standard JSON numbers.

However, numbers over ECMAScript's MAX_SAFE_INTEGER (2⁵³-1) may now
produce JSON that some clients may fail to read correctly, while
previously they would overflow (or error) on the server side.

This shouldn't matter in practice because there shouldn't be such large
numbers for weights, vsizes or mining difficulty. And if there are, the
new behaviour seems (arguably?) better.

To retain DB compatibility, block weights are converted to a u32 for the
BlockMeta storage. This is OK because u32::MAX is far above MAX_BLOCK_WEIGHT.

+++ src/new_index/mempool.rs
@@ -56,7 +56,7 @@ pub struct Mempool {
 pub struct TxOverview {
     txid: Txid,
     fee: u64,
-    vsize: u32,
+    vsize: u64,
     #[cfg(not(feature = "liquid"))]
     value: u64,
 }
@@ -518,9 +518,9 @@ impl Mempool {
 #[derive(Serialize)]
 pub struct BacklogStats {
     pub count: u32,
-    pub vsize: u32,     // in virtual bytes (= weight/4)
+    pub vsize: u64,     // in virtual bytes (= weight/4)
     pub total_fee: u64, // in satoshis
-    pub fee_histogram: Vec<(f32, u32)>,
+    pub fee_histogram: Vec<(f64, u64)>,
 }
 
 impl BacklogStats {
